Expenditure
This is the New Entry page.  The page has three different input fields. The "$" field is a number only page that is inteded to take the user's data. The "Expenditure" takes the name of the purchase. The "Note" field will take any additional information the user wants to input. The calendar will always start on the current date. It will be updated by clicking on the icon. Save will save this expenditure.

The validation and save methods have not been developed.